Correctly split out editor tests namespace
